requirements.txt
asgiref==3.11.0
Django==5.2.9
django-phonenumber-field==8.4.0
django-resized==1.0.3
phonenumbers==9.0.20
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
